GODRIVER-2911: Move OIDC back to auth package, yay
